Created Reusable AppBar
Created a reusable AppBar for main.dart.

AppBar still shows user icon on the right where it should be on the left; the search bar opens the keyboard, but not much else.

main.dart still uses the AppBar regardless of tab; need to create custom widgets for each tab in the BottomNavigationBar.
