+++ Spread/lib/user_widget.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class UserWidget extends StatelessWidget{
+  @override
+  Widget build(BuildContext context) {
+    // TODO: implement build
+    return Container(
+
+    );
+    throw UnimplementedError();
+  }
+
+}
+
